feat: Add playlist collaboration feature This commit introduces a new feature that allows users to collaborate on playlists.
Users can now share their playlists with others, giving them permission to view and modify the content. This required:

Adding a new service to manage who has access to which playlists.

Creating a verification check to ensure users have the right permissions before they can make changes.

+++ src/api/playlists/handler.js
@@ -48,9 +48,9 @@ class PlaylistsHandler {
   async postPlaylistSongHandler(request, h) {
     const { id: playlistId } = request.params;
     const currentUserId = request.auth.credentials.id;
-    await this._playlistsService.verifyPlaylistOwner(playlistId, currentUserId);
     await this._validator.validatePostPlaylistSongPayload(request.payload);
     await this._songsService.verifySong(request.payload.songId);
+    await this._playlistsService.verifyPlaylistAccess(playlistId, currentUserId);
 
     const { songId } = request.payload;
     await this._playlistsService.addPlaylistSong({ playlistId, songId });
@@ -65,6 +65,8 @@ class PlaylistsHandler {
   async getPlaylistSongsHandler(request) {
     const { id: playlistId } = request.params;
     const currentUserId = request.auth.credentials.id;
+    await this._playlistsService.verifyPlaylistAccess(playlistId, currentUserId);
+
     const playlist = await this._playlistsService.getPlaylistSongs(playlistId, currentUserId);
     return {
       status: 'success',
@@ -77,7 +79,8 @@ class PlaylistsHandler {
   async deletePlaylistSongHandler(request) {
     await this._validator.validateDeletePlaylistSongPayload(request.payload);
     const { id: playlistId } = request.params;
-    // const currentUserId = request.auth.credentials.id;
+    const currentUserId = request.auth.credentials.id;
+    await this._playlistsService.verifyPlaylistAccess(playlistId, currentUserId);
 
     const { songId } = request.payload;
     await this._playlistsService.deletePlaylistSong({ playlistId, songId });
